Reorganize navbar files
